sim: deterministic replay pipeline + float64 env + sim eval fixes
Env (Tsimulation/pushshapes/env.py):
- float64 throughout: action_space, obs, step input, _get_obs
- set_state: angle before position (CoG rotation fix)
- get_episode_init(): full init state + obstacles + reset_seed + config_hash

Collection (Tsimulation/collect/):
- zarr_writer: float64 storage for state/actions/goal/reward
- zarr_writer: episode_init JSON attr for deterministic replay
- mouse_collect: pre-step obs, 2x window scale, float64 actions
- scripted_collect: float64 action computation, init_state metadata

Replay (Tsimulation/examples/replay_zarr.py):
- Use episode_init (float64 + reset_seed) when available
- total_frames trim for zero-padding
- --all mode with coverage + drift reporting

Sim eval (egomimic/eval/eval_sim.py):
- float64 for state init and actions
- Fixed reset seed for deterministic replay mode
- Filter batches to target embodiment only
- max_videos support
- pushshapes_sim_stick formatter
- eval_hnet_sim_cotrain.yaml: dual sim eval (circle + stick)

Tests updated: float64 dtype assertions.

Verified: 5/5 scripted + 11/11 mouse episodes replay with zero drift.

+++ Tsimulation/tests/test_features.py
@@ -35,7 +35,7 @@ def test_scripted_action_clipped_and_finite():
         world_size=512.0,
     )
     assert action.shape == (2,)
-    assert action.dtype == np.float32
+    assert action.dtype == np.float64
     assert np.all(action >= 0.0)
     assert np.all(action <= 512.0)
     assert np.all(np.isfinite(action))
